+++ CHANGELOG.md
@@ -37,3 +37,9 @@
 ## 1.0.0+2
 
 - iOS Start meeting issue fixed
+
+## 1.0.0+2
+
+- iOS Start meeting issue fixed
+- iOS get meeting id and password after starting meeting fixed
+- iOS startMeetingNormal added and fixed